requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
youtube-transcript-api
requests
beautifulsoup4
arxiv